+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Fixed
 
 - A bug where the Pill Schedule "First twenty days" mistakenly used "First ten days".
+- Bugs preventing "Last Twenty Days" or "Last Ten Days" from working properly after it's initial use.
+- A bug where "Last Twenty Days" and "Last Ten Days" did one more days than expected.
 - Mistake in Pill Detail view title where it said "New Pill" for the initial pills created.
 
 ### Changed